# Rework and optimize GetLeafData for Light Client