--- a/consensus/mempool_test.go
+++ b/consensus/mempool_test.go
@@ -30,7 +30,6 @@ func TestMempoolNoProgressUntilTxsAvailable(t *testing.T) {
 	config.Consensus.CreateEmptyBlocks = false
 	state, privVals := randGenesisState(1, false, 10)
 	cs := newStateWithConfig(config, state, privVals[0], NewCounterApplication())
-	cs.SetIPFSApi(ipfsTestAPI)
 	assertMempool(cs.txNotifier).EnableTxsAvailable()
 	height, round := cs.Height, cs.Round
 	newBlockCh := subscribe(cs.eventBus, types.EventQueryNewBlock)
@@ -51,7 +50,6 @@ func TestMempoolProgressAfterCreateEmptyBlocksInterval(t *testing.T) {
 	config.Consensus.CreateEmptyBlocksInterval = ensureTimeout
 	state, privVals := randGenesisState(1, false, 10)
 	cs := newStateWithConfig(config, state, privVals[0], NewCounterApplication())
-	cs.SetIPFSApi(ipfsTestAPI)
 
 	assertMempool(cs.txNotifier).EnableTxsAvailable()
 
@@ -70,7 +68,6 @@ func TestMempoolProgressInHigherRound(t *testing.T) {
 	config.Consensus.CreateEmptyBlocks = false
 	state, privVals := randGenesisState(1, false, 10)
 	cs := newStateWithConfig(config, state, privVals[0], NewCounterApplication())
-	cs.SetIPFSApi(ipfsTestAPI)
 
 	assertMempool(cs.txNotifier).EnableTxsAvailable()
 	height, round := cs.Height, cs.Round
@@ -121,7 +118,6 @@ func TestMempoolTxConcurrentWithCommit(t *testing.T) {
 	stateStore := sm.NewStore(blockDB)
 
 	cs := newStateWithConfigAndBlockStore(config, state, privVals[0], NewCounterApplication(), blockDB)
-	cs.SetIPFSApi(ipfsTestAPI)
 	err := stateStore.Save(state)
 	require.NoError(t, err)
 	newBlockHeaderCh := subscribe(cs.eventBus, types.EventQueryNewBlockHeader)
@@ -148,7 +144,6 @@ func TestMempoolRmBadTx(t *testing.T) {
 
 	stateStore := sm.NewStore(blockDB)
 	cs := newStateWithConfigAndBlockStore(config, state, privVals[0], app, blockDB)
-	cs.SetIPFSApi(ipfsTestAPI)
 	err := stateStore.Save(state)
 	require.NoError(t, err)