--- a/light/client.go
+++ b/light/client.go
@@ -8,8 +8,10 @@ import (
 	"time"
 
 	format "github.com/ipfs/go-ipld-format"
-	dag "github.com/ipfs/go-merkledag"
+	"github.com/ipfs/go-merkledag"
 	coreiface "github.com/ipfs/interface-go-ipfs-core"
+	"github.com/lazyledger/nmt/namespace"
+
 	"github.com/lazyledger/lazyledger-core/libs/log"
 	tmmath "github.com/lazyledger/lazyledger-core/libs/math"
 	tmsync "github.com/lazyledger/lazyledger-core/libs/sync"
@@ -17,7 +19,6 @@ import (
 	"github.com/lazyledger/lazyledger-core/light/store"
 	"github.com/lazyledger/lazyledger-core/p2p/ipld"
 	"github.com/lazyledger/lazyledger-core/types"
-	"github.com/lazyledger/nmt/namespace"
 )
 
 type mode byte
@@ -75,7 +76,7 @@ func DataAvailabilitySampling(numSamples uint32, ipfsAPI coreiface.CoreAPI) Opti
 		c.verificationMode = dataAvailabilitySampling
 		c.numSamples = numSamples
 		c.ipfsCoreAPI = ipfsAPI
-		c.dag = dag.NewSession(context.TODO(), ipfsAPI.Dag())
+		c.dag = merkledag.NewSession(context.TODO(), ipfsAPI.Dag())
 	}
 }