--- a/node/node.go
+++ b/node/node.go
@@ -12,6 +12,7 @@ import (
 	"strings"
 	"time"
 
+	format "github.com/ipfs/go-ipld-format"
 	ipface "github.com/ipfs/interface-go-ipfs-core"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
@@ -395,7 +396,7 @@ func createConsensusReactor(config *cfg.Config,
 	csMetrics *cs.Metrics,
 	waitSync bool,
 	eventBus *types.EventBus,
-	ipfs ipface.CoreAPI,
+	dag format.DAGService,
 	consensusLogger log.Logger) (*cs.Reactor, *cs.State) {
 
 	consensusState := cs.NewState(
@@ -404,10 +405,10 @@ func createConsensusReactor(config *cfg.Config,
 		blockExec,
 		blockStore,
 		mempool,
+		dag,
 		evidencePool,
 		cs.StateMetrics(csMetrics),
 	)
-	consensusState.SetIPFSApi(ipfs)
 	consensusState.SetLogger(consensusLogger)
 	if privValidator != nil {
 		consensusState.SetPrivValidator(privValidator)
@@ -757,7 +758,7 @@ func NewNode(config *cfg.Config,
 	}
 	consensusReactor, consensusState := createConsensusReactor(
 		config, state, blockExec, blockStore, mempool, evidencePool,
-		privValidator, csMetrics, stateSync || fastSync, eventBus, ipfs, consensusLogger,
+		privValidator, csMetrics, stateSync || fastSync, eventBus, ipfs.Dag(), consensusLogger,
 	)
 
 	// Set up state sync reactor, and schedule a sync if requested.