--- a/p2p/ipld/racedetector/is_active.go
+++ b/p2p/ipld/racedetector/is_active.go
@@ -1,7 +1,0 @@
-package racedetector
-
-var raceDetectorActive = false
-
-func IsActive() bool {
-	return raceDetectorActive
-}