--- a/p2p/ipld/racedetector/race_detector_file.go
+++ b/p2p/ipld/racedetector/race_detector_file.go
@@ -1,7 +1,0 @@
-// +build race
-
-package racedetector
-
-func init() {
-	raceDetectorActive = true
-}