--- a/p2p/ipld/validate.go
+++ b/p2p/ipld/validate.go
@@ -32,7 +32,6 @@ func ValidateAvailability(
 	numSamples int,
 	onLeafValidity func(namespace.PrefixedData8),
 ) error {
-	// TODO(@Wondertan): Ensure data is fetched within one DAG session
 	ctx, cancel := context.WithTimeout(ctx, ValidationTimeout)
 	defer cancel()