@liamsi 2021-05-24
This looks good. I don't fully understand what makes the big difference in latency possible (yet). Will do a full review after https://github.com/lazyledger/lazyledger-core/pull/352#discussion_r637774071