@Wondertan 2021-05-26
@liamsi, @evan-forbes, I decided not to close this PR and instead extend it with the requested changes for CoreAPI removal