@Wondertan 2021-05-26
In follow-up PR I will go even further and will propose our custom interface for network DataAvailavility which works over DAG