@liamsi 2021-05-26
> race: limit on 8128 simultaneously alive goroutines is exceeded, dying
> FAIL	github.com/lazyledger/lazyledger-core/p2p/ipld	16.744s

This is why the race detector hack was there btw.