@Wondertan 2021-05-26
@liamsi, I guessed that Mock IPFS will decrease the number of goroutines to remove the hack. WIll try something else to fix that