@evan-forbes 2021-05-26
we don't have to merge this tonight. I'll try limiting the goroutines with this PR and see what happens